+++ tests/integration/multidevice/multidevice_component_manager.py
@@ -45,15 +45,6 @@ class MultiDeviceComponentManager(TaskExecutorComponentManager):
 
         super().__init__(logger, *args, matrix_operation="", **kwargs)
 
-    @property
-    def max_executing_tasks(self) -> int:
-        """
-        Get the max number of tasks that can be executing at once.
-
-        :return: max number of simultaneously executing tasks.
-        """
-        return 2
-
     def _non_aborting_lrc(
         self: MultiDeviceComponentManager,
         sleep_time: float,